+++ src/main/java/es/in2/issuer/backend/backoffice/domain/exception/CredentialSerializationException.java
@@ -1,0 +1,7 @@
+package es.in2.issuer.backend.backoffice.domain.exception;
+
+public class CredentialSerializationException extends RuntimeException {
+    public CredentialSerializationException(String message) {
+        super(message);
+    }
+}