+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v2.3.3](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.3.3)
 ### Fixed
 - LEARCredentialMachine mandator validation.
+- In the email sent upon LEARCredentialMachine issuance, use mandator name instead of default.
 
 ## [v2.3.2](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.3.2)
 ### Fixed