+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [v2.3.3](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.3.3)
+### Added
+- Specific exception for credential serialization error.
+
 ## [v2.3.2](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.3.2)
 ### Fixed
 - Add credential status LEARCredentialMachine issuance.